Fire the last 'update' before 'complete' option 2
This merge request is mutually exclusive to https://github.com/juliangarnier/anime/pull/150. Depending on what is desired behavior, either of these should be accepted, the other rejected.
If one should be able to postpone the firing of "complete" in the update callback, this merge request should be accepted. Otherwise https://github.com/juliangarnier/anime/pull/150 should be accepted.