fixes #556: responsive scaling when animating along SVG path
When the svg is scaled by DOM scaling, the path will be scaled twice without this change. This change in function getPathProgress moves away the multiplication of svg dimensions.  When I want to animate obejct along the path(the path is inside a scaled svg), the object will follow the scaled path, which is not what I want. I would like to have the object move along the original path instead of the scaled path.