Update index.html - documentation
Under the documentation for STAGGERING:GRID and STAGGERING:AXIS --- the syntax for `grid: [ rows, columns ]` is incorrect.  This documented syntax is *inverted* relative to anime.js working code. 

The working code for the demo includes `delay: anime.stagger(200, {grid: [14, 5], from: 'center'})`
Of course this correctly animates the 14 col x 5 row grid in the demo.  This pull request corrects the documentation in 3 places - twice in STAGGERING:GRID and once in STAGGERING:AXIS.

Thank you for making such an amazing library, and for this beautiful documentation!