fix: createScope root param type for ReactRef
The `createScope` method accepts `ScopeParams` which allows the `root` property to be
`ReactRef` as part of the union and that type does not accept `null`.

The React libraries type definition for refs is currently:
`type Ref<T> = RefCallback<T> | RefObject<T | null> | null;`

`null` is the valid state for React refs attached to DOM nodes during their initial
render or when unmounting, so if `createScope` should allow passing the ref directly,
as currently shown in the documentation.

For example (from the current docs):
```jsx
// ...
const scope = useRef(null);

useEffect(() => {
  scope.current = createScope({ root })
// ...

return (
  <div ref={root}>
//...
```

This would currently have a typescript error for `root`:
`Type 'null' is not assignable to type 'HTMLElement | SVGElement | undefined'`